pyproject.toml
[project]
name = "ecoagent"
version = "1.0.1"
description = "AI-Powered Sustainability Assistant for Environmental Impact Reduction"
readme = "README.md"
requires-python = ">=3.9"
authors = [
  {name = "EcoAgent Team", email = "team@ecoagent.ai"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.20.0",
    "google-adk>=0.7.0",
    "google-generativeai>=0.8.0",
    "requests>=2.31.0",
    "pydantic>=2.0.0",
    "pydantic[email]>=2.0.0",     # For email validation
    "click>=8.0.0",               # For CLI tools
    "SQLAlchemy>=2.0.0",
    "asyncpg>=0.29.0",
    "python-multipart>=0.0.18",
    "alembic>=1.13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "uv>=0.1.0",
    "pytest-cov>=4.0.0",
    "httpx>=0.27.0",
]

test = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "httpx>=0.27.0",
]

prod = [
    "gunicorn>=20.1.0",
    "sentry-sdk[fastapi]>=2.0.0",
    "prometheus-fastapi-instrumentator>=6.0.0",
    "celery>=5.4.0",
    "redis>=5.0.0",
]

[project.urls]
Homepage = "https://github.com/vinhnx/ecoagent"
Repository = "https://github.com/vinhnx/ecoagent"
Documentation = "https://ecoagent.readthedocs.io"

[project.scripts]
ecoagent = "ecoagent.cli:main"
ecoagent-api = "ecoagent.api:app"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
asyncio_mode = "auto"
addopts = "--strict-markers --strict-config --tb=short"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]

[tool.black]
line-length = 88
target-version = ['py39']

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
plugins = "pydantic.mypy"

[tool.coverage.run]
source = ["ecoagent"]
omit = ["*/tests/*", "*/__pycache__/*", "*/venv/*", "*/.venv/*"]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"